+++ CHANGELOG.md
@@ -26,6 +26,7 @@
   + Fix builder crash when building stores.
   + Fix cache when comparing array values.
   + Stop to pollute global scope.
+  + Optional filter values support specific characters like ' for HTTP requests.
 
 ## 1.4.1
 *2018-09-26*